DB Refactor DBConnectorType Part 1
This refactors DBConnectorType in anticipation of allowing better presentation
of the form fields.  This change moves the config keys out of connectorType.xml
and into DBConnectorType proper, and reorders some of the fields to make more
sense.  It also removes the redundant re-defintion of all the keys,
requiredFields, and configMap from ValidateUtils.

TODO:
  - Add nicer display elements, esp. wrt LOB fields, BaseUrl, etc.
  - Improve layout
  - Make it threadsafe - the ResourceBundle handling is not thread-safe.
  
Code Review:
http://codereview.appspot.com/6554051