Fix issues 4236197 and 5102212: Replace the hostname and dbName configuration properties with googleConnectorName.
The hostname and dbName were used in the display URLs. The dbName was
also used as the root element of the generated XML. The connector name
is now used in both places.

Old: dbconnector://hostname/dbName/docid
New: dbconnector://googleConnectorName.localhost/docid

I preserved the ".localhost" in the display URL because we use it for
googleconnector URLs, and we know the GSA URL handling is OK with that.
I am less sure about how it handles bare hostnames.

This change is not backward compatible with custom stylesheets, since
the root XML element name has (probably) changed. However, this code
was basically pointless, and there are bugs filed against it due to
misunderstandings about how these properties were used. Given the
large-scale changes in the 3.2 release, this seems like a good time to
make this change.

Code review: https://codereview.appspot.com/13582043/